pull helpers out of follower_channel_test.go into helpers_test.go
